test: migrate virtual_attribution.rs unit tests to integration tests
Migrated test_virtual_attributions from src/authorship/virtual_attribution.rs
to tests/integration/virtual_attribution_unit.rs, converting from TmpRepo to
TestRepo. The test verifies VirtualAttributions functionality for tracking file
attributions at a base commit.
