+++ tests/integration/main.rs
@@ -111,6 +111,7 @@ mod sync_authorship_types;
 mod tls_native_certs;
 mod utf8_filenames;
 mod virtual_attribution_merge;
+mod virtual_attribution_unit;
 mod windsurf;
 mod worktrees;
 mod wrapper_performance_targets;
